ai/requirements.txt
pytest==7.4.0
python-dotenv==1.0.0
fastapi==0.100.0
uvicorn==0.23.0
gunicorn==20.1.0
langchain==0.0.266
openai==0.27.8
tiktoken==0.4.0
pyyaml==6.0
requests==2.31.0
peewee==3.16.2
psycopg2-binary==2.9.6
pgvector==0.1.8
sse-starlette==1.6.5

